Scripts/condaenv.7n4k1tbt.requirements.txt
numpy
pandas
PyQt5
matplotlib
pyqtgraph
mediapipe
python-time
pyqt5-tools
pyserial
sklearn